requirements.txt
pdfplumber
langchain
langchain-community
streamlit
requests
fastapi
pydantic
langchain-chroma
langchain-groq
python-dotenv
uvicorn
